feat: rename achievements folder
